1)changes in fillgraph method to remove code to handle ip's not part of top id's separately as "other". 2)changed the data type long to float for dataset of the graph, inorder to avoid truncation of real values.
git-svn-id: https://svn.code.sf.net/p/pmgraph/code/pmgraph@232 903098bd-d3be-4b5e-9df1-ec1ba492e1ff

+++ tests/org/aptivate/bmotools/pmgraph/GraphTestBase.java
@@ -52,7 +52,7 @@ abstract class GraphTestBase extends TestCase
 	 * @param rows 
 	 * @param chart
 	 */
-	protected void checkChartData(long values[][], String[] rows,
+	protected void checkChartData(float values[][], String[] rows,
 			JFreeChart chart)
 	{
